requirements.txt
pandas
numpy
nltk
scikit-learn
gensim
matplotlib
seaborn
jupyter
tweepy
python-dotenv
